+++ backend/canisters/user/impl/src/model/user_canister_event_batch.rs
@@ -1,3 +1,4 @@
+use crate::{can_borrow_state, run_regular_jobs};
 use candid::Deserialize;
 use serde::Serialize;
 use timer_job_queues::{TimerJobItem, TimerJobItemGroup};
@@ -13,6 +14,10 @@ pub struct UserCanisterEventBatch {
 
 impl TimerJobItem for UserCanisterEventBatch {
     async fn process(&self) -> Result<(), bool> {
+        if can_borrow_state() {
+            run_regular_jobs();
+        }
+
         let response = user_canister_c2c_client::c2c_notify_user_canister_events(
             self.user_id.into(),
             &user_canister::c2c_notify_user_canister_events::Args {